+++ example/src/forms/AllValidationsForm/AllValidationsForm.vue
@@ -1,3 +1,4 @@
+<script>
 import Form, {
   Input,
   Checkbox,
@@ -44,7 +45,13 @@ export default {
         <h1>All Validations Form</h1>
         <div class="wrapper">
           <div class="form">
-            <Form cancel submit labelPosition="top" handleSubmit={this.handleSubmit}>
+            <Form
+              cancel
+              submit
+              labelPosition="top"
+              buttonsPosition="end"
+              buttonsClassName="sticky-buttons"
+              handleSubmit={this.handleSubmit}>
               <Input
                 formItem
                 name="username"
@@ -142,3 +149,24 @@ export default {
     )
   },
 }
+</script>
+
+<style scoped lang="less">
+.wrapper {
+  padding-bottom: 60px;
+
+  /deep/ .sticky-buttons {
+    height: 60px;
+    position: fixed;
+    right: 0;
+    bottom: 0;
+    left: 0;
+    align-items: center !important;
+    padding-right: 50px;
+    border-top: 1px solid #dadada;
+    border-bottom: 1px solid #dadada;
+    background-color: #ffffff;
+    box-shadow: 0 0 10px 2px rgba(0, 0, 0, 0.2);
+  }
+}
+</style>